flask-app/requirements.txt
absl-py==0.11.0
appnope==0.1.2
argon2-cffi==20.1.0
astroid==2.6.6
astunparse==1.6.3
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bleach==4.0.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
clang==5.0
click==7.1.2
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
idna==2.10
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
isort==5.9.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==7.0.1
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.6.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
platformdirs==2.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.10.0
pylint==2.9.6
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
PyYAML==5.3.1
pyzmq==22.2.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.7.1
Send2Trash==1.8.0
six==1.15.0
tensorboard==2.4.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-cpu==2.4.3
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.11.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.7.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
